fixed bugs

+++ src/main/java/org/apache/commons/text/StrBuilder.java
@@ -154,7 +154,7 @@ public class StrBuilder implements CharSequence, Appendable, Serializable, Build
         @Override
         public long skip(long n) {
             if (pos + n > StrBuilder.this.size()) {
-                n = StrBuilder.this.size() - pos;
+                n = StrBuilder.this.size() - (long) pos;
             }
             if (n < 0) {
                 return 0;
