+++ src/main/java/org/apache/commons/text/TextStringBuilder.java
@@ -156,7 +156,7 @@ public class TextStringBuilder implements CharSequence, Appendable, Serializable
         @Override
         public long skip(long n) {
             if (pos + n > TextStringBuilder.this.size()) {
-                n = TextStringBuilder.this.size() - pos;
+                n = TextStringBuilder.this.size() - (long) pos;
             }
             if (n < 0) {
                 return 0;
